requirements.txt
pyyaml
six
pathlib
